+++ docs/changes.md
@@ -11,6 +11,7 @@
 - Fixed API changes on Dropbox Provider.
 - Fixed [#15](https://github.com/codenitive/laravel-oneauth/pull/15) where OAuth*\Token\Access return incomplete PHP class from database session, by serialize it first we should avoid such issue and at the same time enable this instance to be reuse to retrieve or send to the provider.
 - Fixed [#16](https://github.com/codenitive/laravel-oneauth/pull/16) Access and Refresh token string length over 255. Increase access token, refresh token and secret length to 500.
+- Add Basecamp OAuth2 Provider.
 
 <a name="0.1.1"></a>
 ## OneAuth 0.1.1